fix(tui): hash message content for chat render cache key
The chat render cache used the byte length of entry.content as its
invalidation key. Two different messages with the same byte length would
collide, so the UI could show stale rendered lines after a message body
changed (e.g. a tool summary or tool_output updated).

- Change ChatRenderCache::entries key from usize length to u64 content
  hash, including both content and tool_output in the hash.
- Update render_chat to compute the hash once per entry and compare it
  against the cached key.

***REMOVED***

+++ src/tui/app.rs
@@ -28,10 +28,10 @@ pub enum ConnectionState {
 /// Cached rendered lines for the chat panel.
 ///
 /// `entries` stores one entry per message in `AppState::messages`. Each entry
-/// records the message content length at the time it was rendered and the
-/// resulting `Line`s (header + body, not the trailing blank separator). The
-/// cache is invalidated when rendering parameters change: terminal width,
-/// search query, or tool-collapse state.
+/// records a stable hash of the message content at the time it was rendered
+/// and the resulting `Line`s (header + body, not the trailing blank
+/// separator). The cache is invalidated when rendering parameters change:
+/// terminal width, search query, or tool-collapse state.
 #[derive(Debug, Default)]
 pub struct ChatRenderCache {
     pub content_width: usize,
@@ -39,7 +39,7 @@ pub struct ChatRenderCache {
     pub tool_collapsed: bool,
     pub expanded_tools: HashSet<usize>,
     pub collapsed_messages: HashSet<usize>,
-    pub entries: Vec<(usize, Vec<Line<'static>>)>,
+    pub entries: Vec<(u64, Vec<Line<'static>>)>,
 }
 
 impl ChatRenderCache {
